move compilation cache to TsCompiler
